Only show the non translated content in metabox
Added a few lines of code to only show the current translation and the non translated posts in the metabox. This might be very useful for those who have a lot of content. I tested this on a demo install with 3 languages and on my main website with a lot of content and 2 different languages. 

Only performs one additional query per language.

I saw this in a feature request, and I thought it may be useful. Also what's the point of showing already translated language in that select?